ai-engine/requirements.txt
anthropic==0.25.0
langchain==0.1.20
langchain-anthropic==0.1.11
redis==5.0.4
python-dotenv==1.0.1
httpx==0.27.0
aiohttp==3.9.5
feedparser==6.0.11
asyncpg==0.29.0
